separate reminder commands/queries into separate services
