+++ src/functions/handlers/cron/reminder-poller.js
@@ -1,5 +1,6 @@
 const cron = require('cron');
-const reminderService = require('../../../services/reminder-service.js');
+const reminderCommandService = require('../../../services/reminder-command-service.js');
+const reminderQueryService = require('../../../services/reminder-query-service.js');
 const CONSTANTS = require('../../../utils/constants.js');
 const logger = require('../../../utils/logger.js');
 
@@ -22,23 +23,23 @@ async function sendReminder(client, reminder, isAdvance) {
 const onTick = async (client) => {
     const now = new Date();
 
-    const advanceReminders = await reminderService.findAdvanceDueReminders(now);
+    const advanceReminders = await reminderQueryService.findAdvanceDueReminders(now);
     // not sure if i like this being 2 try/catches
     // but the service has no try/catch and that seems overkill
     for (const reminder of advanceReminders) {
         try {
             await sendReminder(client, reminder, true);
-            await reminderService.markAlertsSent([reminder.id]);
+            await reminderCommandService.markAlertsSent([reminder.id]);
         } catch (err) {
             logger.error('Failed to send advance reminder', { reminderId: reminder.id, err });
         }
     }
 
-    const finalReminders = await reminderService.findFinalDueReminders(now);
+    const finalReminders = await reminderQueryService.findFinalDueReminders(now);
     for (const reminder of finalReminders) {
         try {
             await sendReminder(client, reminder, false);
-            await reminderService.markComplete([reminder.id]);
+            await reminderCommandService.markComplete([reminder.id]);
         } catch (err) {
             logger.error('Failed to send final reminder', { reminderId: reminder.id, err });
         }
